🔧 chore(package.json): add @types/puppeteer dependency 🔧 chore(tsconfig.json): add src/globals-shim.d.ts to files and include 🔧 chore(types/types.d.ts): move types to types folder 🔧 chore(utils/actions.ts): import types from types/types 🔧 chore(utils/preparedRequests.ts): import types from types/types 🔧 chore(utils/roblox.ts): import CurrentUser from types/types
