+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@types/puppeteer": "^7.0.4",
     "axios": "^1.3.5",
     "chalk": "^4",
     "cheerio": "^1.0.0-rc.12",
