+++ tsconfig.json
@@ -1,4 +1,8 @@
 {
+  "ts-node": {
+    "files": true
+  },
+  "files": ["src/globals-shim.d.ts"],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig to read more about this file */
 
@@ -106,5 +110,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["src/global.d.ts", "src/utils/roblox.ts", "src/index.ts"]
+  "include": ["src/globals-shim.d.ts", "src/utils/roblox.ts", "src/index.ts"]
 }
